LayoutTests/platform/win-wk2/platform/win/plugins/windowless-paint-rect-coordinates-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: Success

This tests that the paint rect passed in WM_PAINT has the expected coordinates.
